\name{GOGraphWrapper}
\alias{GOGraphWrapper}
\title{
Simple wrapper for GOGraph
}
\description{
Utility wrapper
}
\usage{
GOGraphWrapper(term)
}
\arguments{
  \item{term}{
	a GO term
}
}
%\details{
%}
\value{
Result of GOGraph
}
%\references{
%}
\author{
D.Pascovici
}
%\note{
%}

\seealso{
\code{\link{GOGraph}}
}
\examples{

g <- GOGraphWrapper("GO:0009636")

}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
